+++ vue.config.cjs
@@ -3,7 +3,12 @@ const webpack = require('webpack');
 
 module.exports = defineConfig({
   transpileDependencies: true,
-  devServer: {},
+  devServer: {
+    host: '0.0.0.0',
+    port: 8080,
+    open: true,
+    hot: true,
+  },
   lintOnSave: false,
   configureWebpack: {
     plugins: [